feat(avp): AVP.create() as unified API
Now we can simply call AVP.create(int type, ? value) in order to create
a new AVP with the correct type and flags.

+++ src/main/java/com/sipgate/sparta/diameter/core/Answer.java
@@ -62,7 +62,7 @@ public abstract class Answer extends Command {
      *
      * @param resultCode The result code to set.
      */
-    public void setResultCode(final int resultCode) {
+    public void setResultCode(final long resultCode) {
         setAVP(AVP.create(DiameterConstants.AVP_RESULT_CODE, resultCode));
     }
 }
