+++ src/main/java/com/sipgate/sparta/diameter/core/mixins/OriginStateAware.java
@@ -18,7 +18,7 @@ public interface OriginStateAware extends DiameterMessage {
      *
      * @param originStateId The origin state identifier to set.
      */
-    default void setOriginStateId(final int originStateId) {
+    default void setOriginStateId(final long originStateId) {
         setAVP(AVP.create(DiameterConstants.AVP_ORIGIN_STATE_ID, originStateId));
     }
